+++ libengine/ey_signature.h
@@ -33,11 +33,11 @@ extern void ey_free_rhs_item_action(struct ey_engine *eng, ey_rhs_item_action_t 
 struct ey_acsm_pattern;
 typedef struct ey_rhs_item
 {
-	unsigned int signature_id;
-	unsigned int rhs_id;
+	unsigned long signature_id;
+	unsigned long rhs_id;
 
-	unsigned int rhs_signature_position;
-	unsigned int rhs_item_position;
+	unsigned long rhs_signature_position;
+	unsigned long rhs_item_position;
 	TAILQ_ENTRY(ey_rhs_item) link;
 	TAILQ_ENTRY(ey_rhs_item) event_link;
 	ey_location_t location;
@@ -56,8 +56,8 @@ extern void ey_free_rhs_item(struct ey_engine *eng, ey_rhs_item_t *item);
 
 typedef struct ey_rhs_signature
 {
-	unsigned int signature_id;
-	unsigned int rhs_signature_position;
+	unsigned long signature_id;
+	unsigned long rhs_signature_position;
 	ey_location_t location;
 	TAILQ_ENTRY(ey_rhs_signature) link;
 	ey_rhs_item_list_t rhs_item_list;
@@ -69,13 +69,13 @@ extern void ey_free_rhs_signature(struct ey_engine *eng, ey_rhs_signature_t *rhs
 
 typedef struct ey_signature
 {
-	unsigned int signature_id;
+	unsigned long signature_id;
 	TAILQ_ENTRY(ey_signature) link;
 	ey_location_t location;
 	ey_rhs_signature_list_t rhs_signature_list;
 }ey_signature_t;
 typedef TAILQ_HEAD(ey_signature_list, ey_signature) ey_signature_list_t;
-extern ey_signature_t *ey_alloc_signature(struct ey_engine *eng, unsigned int id,
+extern ey_signature_t *ey_alloc_signature(struct ey_engine *eng, unsigned long id,
 	ey_location_t *location, ey_rhs_signature_list_t *signature_list);
 extern void ey_free_signature(struct ey_engine *eng, ey_signature_t *signature);
 
@@ -126,7 +126,7 @@ extern void ey_free_signature_file(struct ey_engine *eng, ey_signature_file_t *f
 
 extern int ey_signature_init(struct ey_engine *eng);
 extern void ey_signature_finit(struct ey_engine *eng);
-extern ey_signature_t *ey_find_signature(struct ey_engine *eng, unsigned int id);
+extern ey_signature_t *ey_find_signature(struct ey_engine *eng, unsigned long id);
 extern int ey_insert_signature(struct ey_engine *eng, ey_signature_t *signature);
 
 extern int ey_signature_add_init(struct ey_engine *eng, const char *function,
